fix: unify runtime and web dependency injection
The web UI was a second WebHost with a second container. ContextualIntelligenceApiController
could not resolve ContextualIntelligenceService, and PatternFileService was constructed twice.

The generic host now owns Kestrel on loopback and every controller resolves from the primary
graph. Audio device open moved off EventMappingService construction so building the graph
never touches hardware.

Closes #23
